+++ src/OpenGraphImage.php
@@ -2,10 +2,26 @@
 
 namespace Vormkracht10\LaravelOpenGraphImage;
 
+use Illuminate\View\ComponentAttributeBag;
+
 class OpenGraphImage
 {
-    public function url(...$args)
+    public function transformAttributeBagToArray(ComponentAttributeBag $attributes): array
     {
-        return url()->signedRoute('open-graph-image.file', ...$args);
+        return collect($attributes)->all();
+    }
+
+    public function url(array|ComponentAttributeBag $parameters): string
+    {
+        if ($parameters instanceof ComponentAttributeBag) {
+            $parameters = $this->transformAttributeBagToArray($parameters);
+        }
+
+        $parameters = collect($parameters)
+            ->merge(['.'.config('open-graph-image.image.extension')]) // add image extension to url for twitter compatibility
+            ->all();
+
+        return url()
+            ->signedRoute('open-graph-image.file', $parameters);
     }
 }